Create new feature branch for main menu with six options
This commit marks the creation of a new feature branch called "main-menu" that will add a main menu to the web app. The menu will display six options that are available to any user of the app.

The main menu will be located at the left of the app and will be accessible from any page. It will provide quick access to the most important features of the app and make it easier for users to navigate.

This commit only creates the new branch and does not contain any actual code changes yet.