FEAT: This conditional checks whether a user and password have been found.
The first line checks whether foundUser is truthy. If it is, it means that a user has been found and its usuario property will be logged to the console using console.log(foundUser.usuario).

If foundUser is truthy, the second line checks whether foundPassword is also truthy. If it is, it means that a password has been found and its clave property will be logged to the console using console.log(foundPassword.clave).

If foundUser is truthy but foundPassword is falsy, it means that a user was found but no password was found. In this case, the message "Password not found" will be logged to the console using console.log("Password not found").

If foundUser is falsy, it means that no user was found. In this case, the message "User not found" will be logged to the console using console.log("User not found").

Overall, this conditional is useful for checking whether a user and password combination exists in a data source and responding appropriately depending on whether they were found or not.